fix(execution): sepolia alpha chainspec
