+++ crates/alloy/hardforks/src/devnet_0.rs
@@ -1,0 +1,18 @@
+/// Bedrock base devnet-0 hardfork activation block is 0.
+pub const BASE_DEVNET_0_BEDROCK_BLOCK: u64 = 0;
+/// Regolith base devnet-0 hardfork activation timestamp is 0.
+pub const BASE_DEVNET_0_REGOLITH_TIMESTAMP: u64 = 0;
+/// Canyon base devnet-0 hardfork activation timestamp is 1698436800.
+pub const BASE_DEVNET_0_CANYON_TIMESTAMP: u64 = 1_698_436_800;
+/// Ecotone base devnet-0 hardfork activation timestamp is 1706634000.
+pub const BASE_DEVNET_0_ECOTONE_TIMESTAMP: u64 = 1_706_634_000;
+/// Fjord base devnet-0 hardfork activation timestamp is 1715961600.
+pub const BASE_DEVNET_0_FJORD_TIMESTAMP: u64 = 1_715_961_600;
+/// Granite base devnet-0 hardfork activation timestamp is 1723046400.
+pub const BASE_DEVNET_0_GRANITE_TIMESTAMP: u64 = 1_723_046_400;
+/// Holocene base devnet-0 hardfork activation timestamp is 1731682800.
+pub const BASE_DEVNET_0_HOLOCENE_TIMESTAMP: u64 = 1_731_682_800;
+/// Isthmus base devnet-0 hardfork activation timestamp is 1744300800.
+pub const BASE_DEVNET_0_ISTHMUS_TIMESTAMP: u64 = 1_744_300_800;
+/// Jovian base devnet-0 hardfork activation timestamp is 1762185600.
+pub const BASE_DEVNET_0_JOVIAN_TIMESTAMP: u64 = 1_762_185_600;
