+++ crates/alloy/hardforks/src/lib.rs
@@ -27,3 +27,11 @@ pub use sepolia::{
     BASE_SEPOLIA_FJORD_TIMESTAMP, BASE_SEPOLIA_GRANITE_TIMESTAMP, BASE_SEPOLIA_HOLOCENE_TIMESTAMP,
     BASE_SEPOLIA_ISTHMUS_TIMESTAMP, BASE_SEPOLIA_JOVIAN_TIMESTAMP, BASE_SEPOLIA_REGOLITH_TIMESTAMP,
 };
+
+mod devnet_0;
+pub use devnet_0::{
+    BASE_DEVNET_0_BEDROCK_BLOCK, BASE_DEVNET_0_CANYON_TIMESTAMP, BASE_DEVNET_0_ECOTONE_TIMESTAMP,
+    BASE_DEVNET_0_FJORD_TIMESTAMP, BASE_DEVNET_0_GRANITE_TIMESTAMP,
+    BASE_DEVNET_0_HOLOCENE_TIMESTAMP, BASE_DEVNET_0_ISTHMUS_TIMESTAMP,
+    BASE_DEVNET_0_JOVIAN_TIMESTAMP, BASE_DEVNET_0_REGOLITH_TIMESTAMP,
+};
