+++ crates/execution/hardforks/src/chain.rs
@@ -76,3 +76,7 @@ pub static BASE_SEPOLIA_HARDFORKS: Lazy<ChainHardforks> =
 /// Base mainnet list of hardforks.
 pub static BASE_MAINNET_HARDFORKS: Lazy<ChainHardforks> =
     Lazy::new(|| OpChainHardforks::base_mainnet().to_chain_hardforks());
+
+/// Base devnet-0 list of hardforks.
+pub static BASE_DEVNET_0_HARDFORKS: Lazy<ChainHardforks> =
+    Lazy::new(|| OpChainHardforks::base_devnet_0().to_chain_hardforks());
